test: split test_commands.py and relocate misplaced tests
Reorganize the test suite for navigability ahead of the stable release
(structural moves only; no test logic changes):

- Split the 2264-line test_commands.py along its 9 class seams into
  test_{preprocessing,reference,segmentation,call,analysis,plots,export,
  batch}.py. test_commands.py becomes a parametrized cnvlib-submodule
  import smoke test (the cheap check that previously rode along).
- Relocate tests to the module they exercise: cluster tests -> ClusterTests
  in test_reference.py; the test_cnvlib OtherTests grab-bag -> test_segfilters.py
  (squash/bic) and test_segmentation.py (transfer_fields); VATests ->
  test_vary.py; CNAryByGeneTests -> test_cnvlib.py; gene_coords_by_name ->
  test_plots.py.
- Centralize the matplotlib "Agg" backend in conftest.py so individual test
  modules no longer set it.
- Drop a duplicate SerialPool test while moving BatchTests (covered in
  test_parallel.py).

Test-only; no changes to cnvlib/ or skgenome/.
