+++ test/test_commands.py
@@ -1,2264 +1,50 @@
 #!/usr/bin/env python
-"""Unit tests for the CNVkit library, cnvlib."""
+"""Smoke test: every cnvlib submodule imports cleanly.
 
-import logging
-import os
-import shutil
-import tempfile
-import unittest
+The per-command unit tests now live in test_<command>.py files; this module
+keeps the cheap import-time smoke check that previously rode along with them.
+"""
+
+import importlib
 
 import pytest
 
-logging.basicConfig(level=logging.ERROR, format="%(message)s")
-
-import warnings
-
-warnings.filterwarnings("ignore", category=DeprecationWarning)
-
-import matplotlib
-import numpy as np
-import pandas as pd
-import pysam
-from skgenome import tabio, GenomicArray as GA
-
-matplotlib.use("Agg")
-
-import cnvlib
-from conftest import linecount
-
-# Import all modules as a smoke test
-from cnvlib import (
-    access,
-    antitarget,
-    autobin,
-    batch,
-    bintest,
-    call,
-    cluster,
-    cmdutil,
-    cnary,
-    commands,
-    core,
-    coverage,
-    diagram,
-    export,
-    fix,
-    heatmap,
-    import_rna,
-    importers,
-    metrics,
-    parallel,
-    params,
-    plots,
-    purity,
-    reference,
-    reports,
-    samutil,
-    scatter,
-    segfilters,
-    segmentation,
-    segmetrics,
-    smoothing,
-    vary,
-)
-
-
-class PreprocessingTests(unittest.TestCase):
-    """Tests for preprocessing commands: access, antitarget, autobin, target, coverage."""
-
-    def test_access(self):
-        fasta = "formats/chrM-Y-trunc.hg19.fa"
-        for min_gap_size, expect_nrows in ((None, 7), (500, 3), (1000, 2)):
-            with self.subTest(min_gap_size=min_gap_size):
-                acc = commands.do_access(
-                    fasta, [], min_gap_size, skip_noncanonical=False
-                )
-                self.assertEqual(len(acc), expect_nrows)
-        excludes = ["formats/dac-my.bed", "formats/my-targets.bed"]
-        for min_gap_size, expect_nrows in (
-            (None, 12),
-            (2, 10),
-            (20, 5),
-            (200, 3),
-            (2000, 2),
-        ):
-            with self.subTest(min_gap_size=min_gap_size, excludes=True):
-                acc = commands.do_access(
-                    fasta, excludes, min_gap_size, skip_noncanonical=False
-                )
-                self.assertEqual(len(acc), expect_nrows)
-        # Dropping chrM, keeping only chrY
-        acc = commands.do_access(fasta, excludes, 10, skip_noncanonical=True)
-        self.assertEqual(len(acc), 5)
-
-    @pytest.mark.slow
-    def test_antitarget(self):
-        """The 'antitarget' command."""
-        baits = tabio.read_auto("formats/nv2_baits.interval_list")
-        acs = tabio.read_auto("../data/access-5k-mappable.hg19.bed")
-        self.assertLess(0, len(commands.do_antitarget(baits)))
-        self.assertLess(0, len(commands.do_antitarget(baits, acs)))
-        self.assertLess(0, len(commands.do_antitarget(baits, acs, 200000)))
-        self.assertLess(0, len(commands.do_antitarget(baits, acs, 10000, 5000)))
-
-    def test_autobin(self):
-        """The 'autobin' command."""
-        bam_fname = "formats/na12878-chrM-Y-trunc.bam"
-        target_bed = "formats/my-targets.bed"
-        targets = tabio.read(target_bed, "bed")
-        access_bed = "../data/access-5k-mappable.hg19.bed"
-        accessible = tabio.read(access_bed, "bed").filter(chromosome="chrY")
-        for method in ("amplicon", "wgs", "hybrid"):
-            (cov, bs), _ = autobin.do_autobin(
-                bam_fname, method, targets=targets, access=accessible
-            )
-            self.assertGreater(cov, 0)
-            self.assertGreater(bs, 0)
-
-    def test_autobin_chrom_name_mismatch(self):
-        """WGS autobin gives a clear, actionable error -- not the cryptic
-        'cannot convert float NaN to integer' -- when the BAM and access
-        regions share no chromosome names (#421)."""
-        bam_fname = "formats/na12878-chrM-Y-trunc.bam"
-        # Access regions on a contig absent from the BAM => no shared chroms,
-        # so the estimated read depth is NaN.
-        access_arr = GA.from_columns(
-            {"chromosome": ["nonexistent_contig"], "start": [0], "end": [100000]}
-        )
-        with self.assertRaises(ValueError) as cm:
-            autobin.do_autobin(bam_fname, "wgs", access=access_arr)
-        msg = str(cm.exception)
-        self.assertNotIn("NaN", msg)
-        self.assertIn("chromosome", msg.lower())
-
-    def test_bedcov_filters_absent_contigs(self):
-        """bedcov keeps regions on BAM contigs and drops those on absent
-        contigs instead of erroring out entirely (#620)."""
-        bam = "formats/na12878-chrM-Y-trunc.bam"
-        bam_chroms = samutil.get_bam_chroms(bam)
-        with tempfile.NamedTemporaryFile("w+t", suffix=".bed", delete=False) as f:
-            f.write("chrM\t251\t277\tfoo\n")
-            f.write("absent_contig\t100\t200\tbar\n")
-            bed = f.name
-        try:
-            table = coverage.bedcov(bed, bam, 0, bam_chroms=bam_chroms)
-            self.assertEqual(list(table.chromosome.unique()), ["chrM"])
-            self.assertEqual(len(table), 1)
-        finally:
-            os.unlink(bed)
-
-    def test_bedcov_all_absent_contigs(self):
-        """bedcov on an all-absent BED: raise a clear error by default, but
-        return empty when allow_empty=True (the per-chunk parallel path)."""
-        bam = "formats/na12878-chrM-Y-trunc.bam"
-        bam_chroms = samutil.get_bam_chroms(bam)
-        with tempfile.NamedTemporaryFile("w+t", suffix=".bed", delete=False) as f:
-            f.write("absent_contig\t100\t200\tbar\n")
-            bed = f.name
-        try:
-            with self.assertRaises(ValueError) as cm:
-                coverage.bedcov(bed, bam, 0, bam_chroms=bam_chroms)
-            self.assertIn("don't match", str(cm.exception))
-            empty = coverage.bedcov(
-                bed, bam, 0, bam_chroms=bam_chroms, allow_empty=True
-            )
-            self.assertEqual(len(empty), 0)
-        finally:
-            os.unlink(bed)
-
-    def test_coverage_skips_marked_duplicates(self):
-        """Coverage ignores reads flagged as duplicates in both the depth
-        (bedcov) and by-count paths, so marking duplicates (e.g. Picard
-        MarkDuplicates) is equivalent to physically removing them (#689)."""
-        seqlen = 50
-        n_total, n_dup = 10, 6  # 4 reads survive the duplicate filter
-        header = {
-            "HD": {"VN": "1.6", "SO": "coordinate"},
-            "SQ": [{"SN": "chr1", "LN": 1000}],
-        }
-        tmpdir = tempfile.mkdtemp()
-        bam = os.path.join(tmpdir, "dup.bam")
-        with pysam.AlignmentFile(bam, "wb", header=header) as out:
-            for i in range(n_total):
-                a = pysam.AlignedSegment()
-                a.query_name = f"r{i}"
-                a.query_sequence = "A" * seqlen
-                a.flag = 0
-                a.reference_id = 0
-                a.reference_start = 100
-                a.mapping_quality = 60
-                a.cigarstring = f"{seqlen}M"
-                a.query_qualities = pysam.qualitystring_to_array("I" * seqlen)
-                a.is_duplicate = i >= (n_total - n_dup)
-                out.write(a)
-        pysam.index(bam)
-        bed = os.path.join(tmpdir, "region.bed")
-        with open(bed, "w") as fh:
-            fh.write("chr1\t90\t160\tregionA\n")
-        n_kept = n_total - n_dup
-        try:
-            # Depth path (default): basecount counts only non-duplicate reads
-            table = coverage.bedcov(bed, bam, 0)
-            self.assertEqual(float(table["basecount"].iloc[0]), n_kept * seqlen)
-            # By-count path: same surviving read count
-            bamfile = pysam.AlignmentFile(bam, "rb")
-            count, _row = coverage.region_depth_count(
-                bamfile, "chr1", 90, 160, "regionA", 0
-            )
-            self.assertEqual(count, n_kept)
-        finally:
-            shutil.rmtree(tmpdir)
-
-    def test_coverage_partial_chrom_mismatch(self):
-        """coverage tolerates a BED mixing present and absent contigs,
-        returning only the present-contig regions, for any process count and
-        either depth/count method (#620: failure was process-count- and
-        method-dependent)."""
-        bam = "formats/na12878-chrM-Y-trunc.bam"
-        with tempfile.NamedTemporaryFile("w+t", suffix=".bed", delete=False) as f:
-            f.write("chrM\t251\t277\tfoo\n")
-            f.write("chrY\t11170\t11213\tbar\n")
-            f.write("absent_contig\t100\t200\tbaz\n")
-            bed = f.name
-        try:
-            for by_count in (False, True):
-                for nprocs in (1, 2):
-                    with self.subTest(by_count=by_count, processes=nprocs):
-                        cna = commands.do_coverage(
-                            bed, bam, by_count=by_count, processes=nprocs
-                        )
-                        self.assertEqual(
-                            sorted(cna.chromosome.unique()), ["chrM", "chrY"]
-                        )
-                        self.assertEqual(len(cna), 2)
-        finally:
-            os.unlink(bed)
-
-    def test_coverage_all_chrom_mismatch(self):
-        """coverage raises a clear error when no BED chromosome matches the
-        BAM, for any process count or method (#620)."""
-        bam = "formats/na12878-chrM-Y-trunc.bam"
-        with tempfile.NamedTemporaryFile("w+t", suffix=".bed", delete=False) as f:
-            f.write("absent_contig\t100\t200\tbaz\n")
-            bed = f.name
-        try:
-            for by_count in (False, True):
-                for nprocs in (1, 2):
-                    with self.subTest(by_count=by_count, processes=nprocs):
-                        with self.assertRaises(ValueError) as cm:
-                            commands.do_coverage(
-                                bed, bam, by_count=by_count, processes=nprocs
-                            )
-                        self.assertIn("don't match", str(cm.exception))
-        finally:
-            os.unlink(bed)
-
-    @pytest.mark.slow
-    def test_coverage(self):
-        """The 'coverage' command."""
-        bed = "formats/my-targets.bed"
-        bam = "formats/na12878-chrM-Y-trunc.bam"
-        # Cover distinct code paths: depth-based default, count-based with
-        # mapq filtering, and multiprocessing.
-        for by_count, min_mapq, nprocs in [
-            (False, 0, 1),
-            (True, 30, 1),
-            (False, 0, 2),
-        ]:
-            cna = commands.do_coverage(
-                bed, bam, by_count=by_count, min_mapq=min_mapq, processes=nprocs
-            )
-            self.assertEqual(len(cna), 4)
-            self.assertTrue((cna.log2 != 0).any())
-            self.assertGreater(cna.log2.nunique(), 1)
-
-    def test_coverage_cram(self):
-        """The 'coverage' command with CRAM input."""
-        bed = "formats/my-targets.bed"
-        cram = "formats/na12878-chrM-Y-trunc.cram"
-        fasta = "formats/chrM-Y-trunc.hg19.fa"
-        bam = "formats/na12878-chrM-Y-trunc.bam"
-        # Verify CRAM coverage matches BAM coverage
-        cna_cram = commands.do_coverage(bed, cram, fasta=fasta)
-        cna_bam = commands.do_coverage(bed, bam)
-        self.assertEqual(len(cna_cram), len(cna_bam))
-        np.testing.assert_array_almost_equal(
-            cna_cram["depth"].values, cna_bam["depth"].values, decimal=1
-        )
-
-    def test_coverage_bedgraph(self):
-        """The 'coverage' command with bedGraph input."""
-        bed = "formats/my-targets.bed"
-        bedgraph = "formats/na12878-chrM-Y-trunc.bed.gz"
-
-        # Test basic bedGraph coverage
-        cna = commands.do_coverage(bed, bedgraph)
-        self.assertEqual(len(cna), 4)
-        self.assertTrue((cna.log2 != 0).any())
-        self.assertGreater(cna.log2.nunique(), 1)
-
-        # Verify sample_id is set correctly
-        self.assertEqual(cna.sample_id, "na12878-chrM-Y-trunc")
-
-        # Verify depth values are calculated
-        self.assertTrue((cna["depth"] >= 0).all())
-
-    def test_coverage_bedgraph_vs_bam(self):
-        """Compare bedGraph and BAM coverage outputs."""
-        bed = "formats/my-targets.bed"
-        bam = "formats/na12878-chrM-Y-trunc.bam"
-        bedgraph = "formats/na12878-chrM-Y-trunc.bed.gz"
-
-        # Get coverage from both inputs
-        cna_bam = commands.do_coverage(bed, bam)
-        cna_bedgraph = commands.do_coverage(bed, bedgraph)
-
-        # Should have same number of regions
-        self.assertEqual(len(cna_bam), len(cna_bedgraph))
-
-        # Convert to DataFrames for easier comparison
-        df_bam = cna_bam.data.copy()
-        df_bedgraph = cna_bedgraph.data.copy()
-
-        # Sort both by chromosome, start, end
-        df_bam = df_bam.sort_values(["chromosome", "start", "end"]).reset_index(
-            drop=True
-        )
-        df_bedgraph = df_bedgraph.sort_values(
-            ["chromosome", "start", "end"]
-        ).reset_index(drop=True)
-
-        # Chromosomes and positions should match exactly
-        pd.testing.assert_series_equal(
-            df_bam["chromosome"], df_bedgraph["chromosome"], check_names=False
-        )
-        pd.testing.assert_series_equal(
-            df_bam["start"], df_bedgraph["start"], check_names=False
-        )
-        pd.testing.assert_series_equal(
-            df_bam["end"], df_bedgraph["end"], check_names=False
-        )
-
-        # Depth values should be reasonably close
-        # Note: Small differences expected due to different depth calculation methods
-        # (pysam.bedcov vs bedtools genomecov)
-        np.testing.assert_allclose(
-            df_bam["depth"].values,
-            df_bedgraph["depth"].values,
-            rtol=0.05,  # Allow 5% relative difference
-            atol=1.0,  # Allow 1.0 absolute difference for low-depth regions
-            err_msg="bedGraph and BAM coverage depths should be reasonably close",
-        )
-
-    def test_coverage_bedgraph_missing_index(self):
-        """Test error handling when tabix index is missing."""
-        bed = "formats/my-targets.bed"
-
-        # Create a temporary bedGraph without index
-        with tempfile.NamedTemporaryFile(suffix=".bed.gz", delete=False) as tmp:
-            tmp_path = tmp.name
-            # Copy bedGraph content but not the index
-            with open("formats/na12878-chrM-Y-trunc.bed.gz", "rb") as src:
-                tmp.write(src.read())
-
-        try:
-            # Should raise FileNotFoundError for missing index
-            with self.assertRaises(FileNotFoundError) as cm:
-                commands.do_coverage(bed, tmp_path)
-
-            self.assertIn("tabix index", str(cm.exception))
-        finally:
-            os.unlink(tmp_path)
-
-    def test_coverage_bedgraph_chr_naming(self):
-        """Test chromosome name mismatch handling (chr1 vs 1)."""
-        bed = "formats/my-targets.bed"
-        # This bedGraph has chromosomes named without 'chr' prefix (M, Y)
-        # while the BED file uses 'chr' prefix (chrM, chrY)
-        bedgraph_nochr = "formats/na12878-M-Y-trunc-nochr.bed.gz"
-
-        # Should still work by auto-matching chromosome names
-        cna = commands.do_coverage(bed, bedgraph_nochr)
-
-        # Should get same number of regions
-        self.assertEqual(len(cna), 4)
-
-        # Should have valid coverage values
-        self.assertTrue((cna.log2 != 0).any())
-        self.assertGreater(cna.log2.nunique(), 1)
-
-    @pytest.mark.slow
-    def test_target_annotate_refflat_assigns_genes(self):
-        """'target --annotate refFlat' actually populates gene names (#688).
-
-        The refFlat reader parses the gene column, tabio sorts the annotation,
-        and into_ranges overlaps it onto the baits. Guards against silent
-        annotation failures: a real bait set should come back almost entirely
-        gene-named, with names drawn from the refFlat's first column.
-        """
-        annot_fname = "formats/refflat-mini.txt"
-        baits = tabio.read_auto("formats/baits-funky.bed")
-        out = commands.do_target(baits, do_split=True, avg_size=200, annotate=annot_fname)
-        named = [g for g in out["gene"] if g != "-"]
-        # Nearly all bins on these gene-dense baits should be annotated
-        self.assertGreater(len(named), 0.9 * len(out))
-        # Names come from the refFlat first column (gene symbols)
-        all_names = {n for g in named for n in g.split(",")}
-        self.assertIn("DDX11L1", all_names)
-        self.assertIn("WASH7P", all_names)
-
-    def test_target_annotate_wrong_build_warns(self):
-        """Warn when annotation shares chrom names but assigns no genes (#688).
-
-        Mimics a wrong-genome-build refFlat: chromosome names match (so
-        compare_chrom_names passes), but no coordinates overlap, leaving every
-        region unnamed. Without the warning this failure is silent.
-        """
-        annot_fname = "formats/refflat-mini.txt"  # chr1 genes end well before 210 Mb
-        baits = GA.from_columns(
-            {
-                "chromosome": ["chr1", "chr1"],
-                "start": [210_000_000, 210_010_000],
-                "end": [210_001_000, 210_011_000],
-                "gene": ["-", "-"],
-            }
-        )
-        with self.assertLogs(level="WARNING") as cm:
-            out = commands.do_target(baits, annotate=annot_fname)
-        self.assertTrue(all(g == "-" for g in out["gene"]))
-        self.assertTrue(any("different genome build" in m for m in cm.output))
-
-    def test_target(self):
-        """The 'target' command."""
-        annot_fname = "formats/refflat-mini.txt"
-        for bait_fname in (
-            "formats/nv2_baits.interval_list",
-            "formats/amplicon.bed",
-            "formats/baits-funky.bed",
-        ):
-            baits = tabio.read_auto(bait_fname)
-            bait_len = len(baits)
-            # No splitting: w/o and w/ re-annotation
-            r1 = commands.do_target(baits)
-            self.assertEqual(len(r1), bait_len)
-            r1a = commands.do_target(baits, do_short_names=True, annotate=annot_fname)
-            self.assertEqual(len(r1a), len(r1))
-            # Splitting, w/o and w/ re-annotation
-            r2 = commands.do_target(
-                baits, do_short_names=True, do_split=True, avg_size=100
-            )
-            self.assertGreater(len(r2), len(r1))
-            for _c, subarr in r2.by_chromosome():
-                self.assertTrue(subarr.start.is_monotonic_increasing, bait_fname)
-                self.assertTrue(subarr.end.is_monotonic_increasing, bait_fname)
-                # Bins are non-overlapping; next start >= prev. end
-                self.assertTrue(
-                    (
-                        (subarr.start.to_numpy()[1:] - subarr.end.to_numpy()[:-1]) >= 0
-                    ).all()
-                )
-            r2a = commands.do_target(
-                baits,
-                do_short_names=True,
-                do_split=True,
-                avg_size=100,
-                annotate=annot_fname,
-            )
-            self.assertEqual(len(r2a), len(r2))
-            # Original regions object should be unmodified
-            self.assertEqual(len(baits), bait_len)
-
-
-class ReferenceTests(unittest.TestCase):
-    """Tests for the reference and fix commands."""
-
-    def test_reference(self):
-        """The 'reference' command."""
-        # Empty/unspecified antitargets
-        nlines = linecount("formats/amplicon.cnr") - 1
-        ref = commands.do_reference(["formats/amplicon.cnr"], ["formats/empty"])
-        self.assertEqual(len(ref), nlines)
-        ref = commands.do_reference(["formats/amplicon.cnr"])
-        self.assertEqual(len(ref), nlines)
-        # Empty/unspecified antitargets, flat reference
-        nlines = linecount("formats/amplicon.bed")
-        ref = commands.do_reference_flat("formats/amplicon.bed", "formats/empty")
-        self.assertEqual(len(ref), nlines)
-        ref = commands.do_reference_flat("formats/amplicon.bed")
-        self.assertEqual(len(ref), nlines)
-        # Misc
-        ref = cnvlib.read("formats/reference-tr.cnn")
-        targets, antitargets = reference.reference2regions(ref)
-        self.assertLess(0, len(antitargets))
-        self.assertEqual(len(antitargets), (ref["gene"] == "Background").sum())
-        self.assertEqual(len(targets), len(ref) - len(antitargets))
-
-    def test_reference_gender_input(self):
-        """Test whether correct log2-ratios are calculated for sex chromosomes in reference command"""
-        expected_haploid_log2 = [-1.0]
-        expected_diploid_log2 = [0.0]
-
-        # Test case when sample gender is provided by user
-        ref_male = commands.do_reference(
-            ["formats/ref_test_male.cnn"], female_samples=False
-        )  # is_haploid_x_reference=False
-        ref_female = commands.do_reference(
-            ["formats/ref_test_female.cnn"], female_samples=True
-        )  # is_haploid_x_reference=False
-        self.assertListEqual(
-            ref_male["log2"][ref_male.chr_x_filter()].to_list(), expected_diploid_log2
-        )
-        self.assertListEqual(
-            ref_male["log2"][ref_male.chr_y_filter()].to_list(), expected_haploid_log2
-        )
-        self.assertListEqual(
-            ref_female["log2"][ref_female.chr_x_filter()].to_list(),
-            expected_diploid_log2,
-        )
-        ref_male = commands.do_reference(
-            ["formats/ref_test_male.cnn"],
-            female_samples=False,
-            is_haploid_x_reference=True,
-        )
-        ref_female = commands.do_reference(
-            ["formats/ref_test_female.cnn"],
-            female_samples=True,
-            is_haploid_x_reference=True,
-        )
-        self.assertListEqual(
-            ref_male["log2"][ref_male.chr_x_filter()].to_list(), expected_haploid_log2
-        )
-        self.assertListEqual(
-            ref_male["log2"][ref_male.chr_y_filter()].to_list(), expected_haploid_log2
-        )
-        self.assertListEqual(
-            ref_female["log2"][ref_female.chr_x_filter()].to_list(),
-            expected_haploid_log2,
-        )
-
-        # Test case when sample gender is guessed from input
-        ref_male = commands.do_reference(
-            ["formats/ref_test_male.cnn"]
-        )  # is_haploid_x_reference=False
-        ref_female = commands.do_reference(
-            ["formats/ref_test_female.cnn"]
-        )  # is_haploid_x_reference=False
-        self.assertListEqual(
-            ref_male["log2"][ref_male.chr_x_filter()].to_list(), expected_diploid_log2
-        )
-        self.assertListEqual(
-            ref_male["log2"][ref_male.chr_y_filter()].to_list(), expected_haploid_log2
-        )
-        self.assertListEqual(
-            ref_female["log2"][ref_female.chr_x_filter()].to_list(),
-            expected_diploid_log2,
-        )
-        ref_male = commands.do_reference(
-            ["formats/ref_test_male.cnn"], is_haploid_x_reference=True
-        )
-        ref_female = commands.do_reference(
-            ["formats/ref_test_female.cnn"], is_haploid_x_reference=True
-        )
-        self.assertListEqual(
-            ref_male["log2"][ref_male.chr_x_filter()].to_list(), expected_haploid_log2
-        )
-        self.assertListEqual(
-            ref_male["log2"][ref_male.chr_y_filter()].to_list(), expected_haploid_log2
-        )
-        self.assertListEqual(
-            ref_female["log2"][ref_female.chr_x_filter()].to_list(),
-            expected_haploid_log2,
-        )
-
-    def test_cluster_hierarchical(self):
-        """Test hierarchical clustering on synthetic data with 2 clear groups."""
-        from cnvlib.cluster import hierarchical
-
-        rng = np.random.default_rng(42)
-        n_bins = 100
-        # Group A: 5 samples with similar pattern
-        base_a = rng.standard_normal(n_bins)
-        group_a = np.array([base_a + rng.normal(0, 0.1, n_bins) for _ in range(5)])
-        # Group B: 5 samples with a different pattern
-        base_b = rng.standard_normal(n_bins)
-        group_b = np.array([base_b + rng.normal(0, 0.1, n_bins) for _ in range(5)])
-        samples = np.vstack([group_a, group_b])
-
-        clusters = hierarchical(samples, min_cluster_size=3)
-        # Should produce exactly 2 clusters; all samples must be assigned
-        self.assertEqual(len(clusters), 2)
-        all_indices = sorted(idx for c in clusters for idx in c)
-        self.assertEqual(all_indices, list(range(10)))
-        # Check they separate the groups correctly
-        c0 = set(clusters[0])
-        c1 = set(clusters[1])
-        group_a_set = set(range(5))
-        group_b_set = set(range(5, 10))
-        self.assertTrue(
-            (c0 == group_a_set and c1 == group_b_set)
-            or (c0 == group_b_set and c1 == group_a_set)
-        )
-
-    def test_cluster_kmedoids(self):
-        """Test k-medoids clustering produces valid clusters."""
-        from cnvlib.cluster import kmedoids
-
-        rng = np.random.default_rng(42)
-        n_bins = 100
-        samples = np.vstack(
-            [
-                rng.standard_normal((5, n_bins)),
-                rng.standard_normal((5, n_bins)) + 3,
-            ]
-        )
-        clusters = kmedoids(samples, k=2)
-        self.assertEqual(len(clusters), 2)
-        all_indices = sorted(idx for c in clusters for idx in c)
-        self.assertEqual(all_indices, list(range(10)))
-
-    def test_create_clusters_column_names(self):
-        """Test that create_clusters produces expected column names."""
-        rng = np.random.default_rng(42)
-        n_bins = 50
-        n_samples = 6
-        # Create logr_matrix with pseudocount row + sample rows
-        logr = np.vstack(
-            [
-                np.zeros(n_bins),  # pseudocount
-                rng.standard_normal((n_samples, n_bins)),
-            ]
-        )
-        depths = np.vstack(
-            [
-                np.zeros(n_bins),  # pseudocount
-                rng.uniform(10, 100, (n_samples, n_bins)),
-            ]
-        )
-        sample_ids = [f"sample_{i}" for i in range(n_samples)]
-        cols = reference.create_clusters(
-            logr, depths, min_cluster_size=2, sample_ids=sample_ids
-        )
-        # Should have at least one cluster with log2_, depth_, spread_ columns
-        self.assertTrue(len(cols) > 0)
-        log2_keys = [k for k in cols if k.startswith("log2_")]
-        depth_keys = [k for k in cols if k.startswith("depth_")]
-        spread_keys = [k for k in cols if k.startswith("spread_")]
-        self.assertTrue(len(log2_keys) > 0)
-        self.assertEqual(len(log2_keys), len(depth_keys))
-        self.assertEqual(len(log2_keys), len(spread_keys))
-        # All column arrays should have n_bins elements
-        for val in cols.values():
-            self.assertEqual(len(val), n_bins)
-
-    def test_fix(self):
-        """The 'fix' command."""
-        # Extract fake target/antitarget bins from a combined file
-        ref = cnvlib.read("formats/reference-tr.cnn")
-        is_bg = ref["gene"] == "Background"
-        tgt_bins = ref[~is_bg]
-        rng = np.random.default_rng(42)  # Use fixed seed for reproducible tests
-        tgt_bins.log2 += rng.standard_normal(len(tgt_bins)) / 5
-        anti_bins = ref[is_bg]
-        anti_bins.log2 += rng.standard_normal(len(anti_bins)) / 5
-        blank_bins = cnary.CopyNumArray([])
-        # Typical usage (hybrid capture)
-        cnr = commands.do_fix(tgt_bins, anti_bins, ref)
-        self.assertTrue(0 < len(cnr) <= len(ref))
-        # Blank antitargets (WGS or amplicon)
-        cnr = commands.do_fix(tgt_bins, blank_bins, ref[~is_bg])
-        self.assertTrue(0 < len(cnr) <= len(tgt_bins))
-
-    def test_fix_degenerate_no_nan(self):
-        """do_fix never emits NaN log2/weight from degenerate input (#521/#524).
-
-        Zero-coverage sentinel and NaN bins (from malformed inputs or dead
-        regions in WGS) must not survive into the .cnr, where they would crash
-        smoothing/CBS downstream.
-        """
-        ref = cnvlib.read("formats/reference-tr.cnn")
-        is_bg = ref["gene"] == "Background"
-        tgt_bins = ref[~is_bg].copy()
-        anti_bins = ref[is_bg].copy()
-        # Inject degenerate values into a few sample bins
-        tgt_log2 = tgt_bins["log2"].copy()
-        tgt_log2.iloc[0] = np.nan
-        tgt_log2.iloc[1] = params.NULL_LOG2_COVERAGE
-        tgt_bins["log2"] = tgt_log2
-        # Inject a degenerate value into the reference, too
-        ref_log2 = ref["log2"].copy()
-        ref_log2.iloc[2] = np.nan
-        ref["log2"] = ref_log2
-        cnr = commands.do_fix(tgt_bins, anti_bins, ref)
-        self.assertFalse(np.isnan(cnr["log2"]).any())
-        self.assertFalse(np.isnan(cnr["weight"]).any())
-
-
-class SegmentationTests(unittest.TestCase):
-    """Tests for segmentation commands."""
-
-    def test_segment(self):
-        """The 'segment' command."""
-        cnarr = cnvlib.read("formats/amplicon.cnr")
-        n_chroms = cnarr.chromosome.nunique()
-        # NB: R methods are in another script; haar is pure-Python
-        segments = segmentation.do_segmentation(cnarr, "haar")
-        self.assertGreater(len(segments), n_chroms)
-        self.assertTrue((segments.start < segments.end).all())
-        segments = segmentation.do_segmentation(
-            cnarr, "haar", threshold=0.0001, skip_low=True
-        )
-        self.assertGreater(len(segments), n_chroms)
-        self.assertTrue((segments.start < segments.end).all())
-        varr = tabio.read("formats/na12878_na12882_mix.vcf", "vcf")
-        # TODO - This test is failing... commenting it out for now!
-        # segments = segmentation.do_segmentation(cnarr, "haar", variants=varr)
-        # self.assertGreater(len(segments), n_chroms)
-        # self.assertTrue((segments.start < segments.end).all())
-
-    @pytest.mark.slow
-    def test_segment_hmm(self):
-        """The 'segment' command with HMM methods."""
-        # Test all HMM method variants on one file
-        cnarr = cnvlib.read("formats/amplicon.cnr")
-        n_chroms = cnarr.chromosome.nunique()
-        # NB: R methods are in another script; haar is pure-Python
-        segments = segmentation.do_segmentation(cnarr, "hmm")
-        self.assertGreater(len(segments), n_chroms)
-        self.assertTrue((segments.start < segments.end).all())
-        segments = segmentation.do_segmentation(cnarr, "hmm-tumor", skip_low=True)
-        self.assertGreater(len(segments), n_chroms)
-        self.assertTrue((segments.start < segments.end).all())
-        segments = segmentation.do_segmentation(cnarr, "hmm-germline")
-        self.assertGreater(len(segments), n_chroms)
-        self.assertTrue((segments.start < segments.end).all())
-        varr = tabio.read("formats/na12878_na12882_mix.vcf", "vcf")
-        segments = segmentation.do_segmentation(cnarr, "hmm", variants=varr)
-        self.assertGreater(len(segments), n_chroms)
-        # Verify default HMM also works on a different dataset
-        cnarr2 = cnvlib.read("formats/p2-20_1.cnr")
-        n_chroms2 = cnarr2.chromosome.nunique()
-        segments = segmentation.do_segmentation(cnarr2, "hmm")
-        self.assertGreater(len(segments), n_chroms2)
-        self.assertTrue((segments.start < segments.end).all())
-
-    @pytest.mark.slow
-    def test_segment_parallel(self):
-        """The 'segment' command, in parallel."""
-        cnarr = cnvlib.read("formats/amplicon.cnr")
-        psegments = segmentation.do_segmentation(cnarr, "haar", processes=2)
-        ssegments = segmentation.do_segmentation(cnarr, "haar", processes=1)
-        self.assertEqual(psegments.data.shape, ssegments.data.shape)
-        self.assertEqual(len(psegments.meta), len(ssegments.meta))
-
-    def test_segment_empty_input(self):
-        """Test segmentation with empty CNR input (issue #970)."""
-        # Create an empty CNA with proper structure (header only)
-        empty_data = pd.DataFrame(
-            columns=["chromosome", "start", "end", "gene", "log2"]
-        )
-        empty_cnarr = cnvlib.cnary.CopyNumArray(empty_data, {"sample_id": "test"})
-
-        # Test with serial processing
-        segments = segmentation.do_segmentation(empty_cnarr, "haar", processes=1)
-        self.assertEqual(len(segments), 0)
-        self.assertListEqual(
-            list(segments.data.columns), list(empty_cnarr.data.columns)
-        )
-
-        # Test with parallel processing
-        psegments = segmentation.do_segmentation(empty_cnarr, "haar", processes=2)
-        self.assertEqual(len(psegments), 0)
-
-        # Test with save_dataframe=True
-        segments_df, rstr = segmentation.do_segmentation(
-            empty_cnarr, "haar", processes=1, save_dataframe=True
-        )
-        self.assertEqual(len(segments_df), 0)
-        self.assertEqual(rstr, "")
-
-
-class CallTests(unittest.TestCase):
-    """Tests for the call command and related functions."""
-
-    def test_call(self):
-        """The 'call' command."""
-        # Methods: clonal, threshold, none
-        tr_cns = cnvlib.read("formats/tr95t.cns")
-        tr_thresh = commands.do_call(
-            tr_cns,
-            None,
-            "threshold",
-            is_haploid_x_reference=True,
-            is_sample_female=True,
-        )
-        self.assertEqual(len(tr_cns), len(tr_thresh))
-        tr_clonal = commands.do_call(
-            tr_cns,
-            None,
-            "clonal",
-            purity=0.65,
-            is_haploid_x_reference=True,
-            is_sample_female=True,
-        )
-        self.assertEqual(len(tr_cns), len(tr_clonal))
-        cl_cns = cnvlib.read("formats/cl_seq.cns")
-        cl_thresh = commands.do_call(
-            cl_cns,
-            None,
-            "threshold",
-            thresholds=np.log2((np.arange(12) + 0.5) / 6.0),
-            is_haploid_x_reference=True,
-            is_sample_female=True,
-        )
-        self.assertEqual(len(cl_cns), len(cl_thresh))
-        cl_clonal = commands.do_call(
-            cl_cns,
-            None,
-            "clonal",
-            ploidy=6,
-            purity=0.99,
-            is_haploid_x_reference=True,
-            is_sample_female=True,
-        )
-        self.assertEqual(len(cl_cns), len(cl_clonal))
-        cl_none = commands.do_call(
-            cl_cns,
-            None,
-            "none",
-            ploidy=6,
-            purity=0.99,
-            is_haploid_x_reference=True,
-            is_sample_female=True,
-        )
-        self.assertEqual(len(cl_cns), len(cl_none))
-
-    def test_call_filter(self):
-        segments = cnvlib.read("formats/tr95t.segmetrics.cns")
-        variants = tabio.read("formats/na12878_na12882_mix.vcf", "vcf")
-        # Each filter individually, then filter combos (with merges where needed)
-        for filters, merges in (
-            (["ampdel"], []),
-            (["cn"], []),
-            (["ci"], []),
-            (["sem"], []),
-            ([], ["bic"]),
-            (["sem", "cn", "ampdel"], []),
-            (["ci", "cn"], []),
-            (["ci", "sem"], []),
-            (["cn", "ci", "sem"], []),
-            ([], ["bic", "cn"]),
-            (["ci"], ["bic"]),
-            (["ci", "cn"], ["bic"]),
-        ):
-            with self.subTest(filters=filters, merges=merges):
-                result = commands.do_call(
-                    segments,
-                    variants,
-                    method="threshold",
-                    purity=0.9,
-                    is_haploid_x_reference=True,
-                    is_sample_female=True,
-                    filters=filters or None,
-                    merges=merges or None,
-                )
-                self.assertLessEqual(len(result), len(segments))
-                if "ampdel" not in filters:
-                    # At least 1 segment per chromosome remains
-                    self.assertLessEqual(len(segments.chromosome.unique()), len(result))
-                for colname in "baf", "cn", "cn1", "cn2":
-                    self.assertIn(colname, result)
-                # Segmetrics columns must survive merging by prior filters
-                for colname in "sem", "ci_lo", "ci_hi", "stdev":
-                    self.assertIn(
-                        colname,
-                        result,
-                        f"{colname!r} dropped after filters={filters} merges={merges}",
-                    )
-
-    def test_call_filter_ci_preserves_different_magnitudes(self):
-        """CI filter should not merge adjacent segments with different magnitudes.
-
-        Two adjacent segments both with CIs entirely below zero (both are
-        confident losses) should remain separate if they have different log2
-        values.  Only segments whose CI overlaps zero (neutral) should be
-        merged with adjacent neutral segments.
-        """
-        segarr = cnary.CopyNumArray(
-            pd.DataFrame(
-                {
-                    "chromosome": ["chr1"] * 5,
-                    "start": [0, 1000, 2000, 3000, 4000],
-                    "end": [1000, 2000, 3000, 4000, 5000],
-                    "gene": ["A", "B", "C", "D", "E"],
-                    "log2": [-1.0, -0.03, 0.01, 0.02, 0.5],
-                    "weight": [1.0, 1.0, 1.0, 1.0, 1.0],
-                    "probes": [10, 10, 10, 10, 10],
-                    # Segment A: deep loss, CI entirely below 0
-                    # Segment B: slight loss, CI entirely below 0
-                    # Segment C: neutral, CI overlaps 0
-                    # Segment D: neutral, CI overlaps 0
-                    # Segment E: gain, CI entirely above 0
-                    "ci_lo": [-1.2, -0.06, -0.05, -0.04, 0.3],
-                    "ci_hi": [-0.8, -0.01, 0.07, 0.08, 0.7],
-                }
-            )
-        )
-
-        result = segfilters.ci(segarr)
-
-        # A and B must stay separate -- both are losses but different magnitude
-        # C and D should merge -- both are neutral (CI overlaps zero)
-        # E must stay separate -- it's a gain
-        self.assertEqual(len(result), 4)
-        # Check that the deep loss is preserved with its own log2
-        self.assertAlmostEqual(result["log2"].iat[0], -1.0)
-        # Check that the slight loss is preserved with its own log2
-        self.assertAlmostEqual(result["log2"].iat[1], -0.03)
-        # Check that the two neutrals were merged
-        self.assertEqual(result["probes"].iat[2], 20)
-        # Check that the gain is preserved
-        self.assertAlmostEqual(result["log2"].iat[3], 0.5)
-
-    def test_call_filter_bic_different_means_stay_separate(self):
-        """BIC filter should not merge segments with clearly different means."""
-        segarr = cnary.CopyNumArray(
-            pd.DataFrame(
-                {
-                    "chromosome": ["chr1"] * 3,
-                    "start": [0, 1000, 2000],
-                    "end": [1000, 2000, 3000],
-                    "gene": ["A", "B", "C"],
-                    "log2": [-1.0, 0.0, 0.5],
-                    "weight": [10.0, 10.0, 10.0],
-                    "probes": [50, 50, 50],
-                    "stdev": [0.1, 0.1, 0.1],
-                }
-            )
-        )
-        result = segfilters.bic(segarr)
-        # All three segments have very different means with low variance
-        self.assertEqual(len(result), 3)
-
-    def test_call_filter_bic_similar_means_merge(self):
-        """BIC filter should merge adjacent segments with similar means."""
-        segarr = cnary.CopyNumArray(
-            pd.DataFrame(
-                {
-                    "chromosome": ["chr1"] * 3,
-                    "start": [0, 1000, 2000],
-                    "end": [1000, 2000, 3000],
-                    "gene": ["A", "B", "C"],
-                    "log2": [0.31, 0.30, 0.29],
-                    "weight": [10.0, 10.0, 10.0],
-                    "probes": [20, 20, 20],
-                    "stdev": [0.5, 0.5, 0.5],
-                }
-            )
-        )
-        result = segfilters.bic(segarr)
-        # All three are essentially the same; BIC should merge them
-        self.assertEqual(len(result), 1)
-        self.assertEqual(result["probes"].iat[0], 60)
-
-    def test_call_filter_bic_chromosome_boundary(self):
-        """BIC filter should never merge segments on different chromosomes."""
-        segarr = cnary.CopyNumArray(
-            pd.DataFrame(
-                {
-                    "chromosome": ["chr1", "chr2"],
-                    "start": [0, 0],
-                    "end": [1000, 1000],
-                    "gene": ["A", "B"],
-                    "log2": [0.3, 0.3],
-                    "weight": [10.0, 10.0],
-                    "probes": [20, 20],
-                    "stdev": [0.5, 0.5],
-                }
-            )
-        )
-        result = segfilters.bic(segarr)
-        # Same means but different chromosomes: must not merge
-        self.assertEqual(len(result), 2)
-
-    def test_call_filter_bic_single_bin_fallback(self):
-        """BIC filter handles segments with stdev=0 via fallback variance."""
-        segarr = cnary.CopyNumArray(
-            pd.DataFrame(
-                {
-                    "chromosome": ["chr1"] * 3,
-                    "start": [0, 1000, 2000],
-                    "end": [1000, 2000, 3000],
-                    "gene": ["A", "B", "C"],
-                    "log2": [0.3, 0.3, 0.3],
-                    "weight": [1.0, 10.0, 10.0],
-                    "probes": [1, 20, 20],
-                    "stdev": [0.0, 0.5, 0.5],
-                }
-            )
-        )
-        result = segfilters.bic(segarr)
-        # Segment A has stdev=0 (single bin), fallback variance used;
-        # all have similar means so they should merge
-        self.assertEqual(len(result), 1)
-        self.assertEqual(result["probes"].iat[0], 41)
-
-    def test_call_filter_bic_cascading_merges(self):
-        """BIC filter iteratively merges: A~B and B~C yields A+B+C."""
-        segarr = cnary.CopyNumArray(
-            pd.DataFrame(
-                {
-                    "chromosome": ["chr1"] * 4,
-                    "start": [0, 1000, 2000, 3000],
-                    "end": [1000, 2000, 3000, 4000],
-                    "gene": ["A", "B", "C", "D"],
-                    "log2": [0.10, 0.12, 0.14, -0.5],
-                    "weight": [10.0, 10.0, 10.0, 10.0],
-                    "probes": [20, 20, 20, 20],
-                    "stdev": [0.4, 0.4, 0.4, 0.1],
-                }
-            )
-        )
-        result = segfilters.bic(segarr)
-        # A, B, C have similar means and high variance -> merge into one
-        # D has a clearly different mean -> stays separate
-        self.assertEqual(len(result), 2)
-        self.assertEqual(result["probes"].iat[0], 60)
-        self.assertEqual(result["probes"].iat[1], 20)
-
-    def test_call_merge_bic_via_do_call(self):
-        """BIC merge works through the do_call pipeline."""
-        segments = cnvlib.read("formats/tr95t.segmetrics.cns")
-        result = commands.do_call(
-            segments,
-            variants=None,
-            method="threshold",
-            purity=0.9,
-            is_haploid_x_reference=True,
-            is_sample_female=True,
-            merges=["bic"],
-        )
-        self.assertGreater(len(result), 0)
-        self.assertLessEqual(len(result), len(segments))
-
-    def test_call_filter_cn_vs_merge_cn(self):
-        """--filter cn merges only neutral segments; --merge cn merges all same-CN."""
-        segarr = cnary.CopyNumArray(
-            pd.DataFrame(
-                {
-                    "chromosome": ["chr1"] * 4,
-                    "start": [0, 1000, 2000, 3000],
-                    "end": [1000, 2000, 3000, 4000],
-                    "gene": ["A", "B", "C", "D"],
-                    "log2": [-0.5, -0.6, 0.0, 0.01],
-                    "weight": [1.0, 1.0, 1.0, 1.0],
-                    "probes": [10, 10, 10, 10],
-                }
-            )
-        )
-        # --filter cn: only merges adjacent neutral (cn==2) segments
-        result_filter = commands.do_call(
-            segarr,
-            method="threshold",
-            filters=["cn"],
-        )
-        # Segments A,B are cn=1 (losses) — NOT merged by --filter cn
-        # Segments C,D are cn=2 (neutral) — merged by --filter cn
-        self.assertEqual(len(result_filter), 3)
-
-        # --merge cn: merges any adjacent same-CN segments
-        result_merge = commands.do_call(
-            segarr,
-            method="threshold",
-            merges=["cn"],
-        )
-        # Segments A,B are cn=1 — merged by --merge cn
-        # Segments C,D are cn=2 — merged by --merge cn
-        self.assertEqual(len(result_merge), 2)
-
-    def test_call_filter_cn_neutral_sex_chrom(self):
-        """--filter cn respects expected CN on sex chromosomes (male sample)."""
-        segarr = cnary.CopyNumArray(
-            pd.DataFrame(
-                {
-                    "chromosome": ["chr1", "chr1", "chrX", "chrX"],
-                    "start": [0, 1000, 0, 1000],
-                    "end": [1000, 2000, 1000, 2000],
-                    "gene": ["A", "B", "C", "D"],
-                    "log2": [0.0, 0.01, 0.0, 0.05],
-                    "weight": [1.0, 1.0, 1.0, 1.0],
-                    "probes": [10, 10, 10, 10],
-                }
-            )
-        )
-        # Male sample: expected CN=2 on autosomes, CN=1 on chrX
-        result = commands.do_call(
-            segarr,
-            method="threshold",
-            is_haploid_x_reference=True,
-            is_sample_female=False,
-            filters=["cn"],
-        )
-        # chr1 A,B are cn=2 (neutral for autosome) — merged
-        # chrX C,D are cn=1 (neutral for male X) — merged
-        self.assertEqual(len(result), 2)
-
-        # Female sample: expected CN=2 on chrX too
-        result_f = commands.do_call(
-            segarr,
-            method="threshold",
-            is_haploid_x_reference=True,
-            is_sample_female=True,
-            filters=["cn"],
-        )
-        # chr1 A,B are cn=2 (neutral) — merged
-        # chrX C,D are cn=1 (NOT neutral for female, expected 2) — kept separate
-        self.assertEqual(len(result_f), 3)
-
-    def test_call_log2_ratios(self):
-        cnarr = cnvlib.read("formats/par-reference.grch38.cnn")
-        ploidy = 2
-        purity = 0.8
-        is_haploid_x_reference = True
-        is_sample_female = False
-        diploid_parx_genome = None
-        absolutes = call.absolute_clonal(
-            cnarr,
-            ploidy,
-            purity,
-            is_haploid_x_reference,
-            diploid_parx_genome,
-            is_sample_female,
-        )
-        ratios = call.log2_ratios(
-            cnarr, absolutes, ploidy, is_haploid_x_reference, diploid_parx_genome
-        )
-        ratios_dprx = call.log2_ratios(
-            cnarr, absolutes, ploidy, is_haploid_x_reference, "grch38"
-        )
-        self.assertEqual(len(ratios), len(cnarr))
-        self.assertEqual(len(ratios_dprx), len(cnarr))
-        self.assertEqual(ratios[26], ratios_dprx[26])
-        self.assertEqual(ratios[26], ratios_dprx[27] + 1)
-
-    def test_call_sex(self):
-        """Test each 'call' method on allosomes."""
-        for (
-            fname,
-            sample_is_f,
-            ref_is_m,
-            chr1_expect,
-            chrx_expect,
-            chry_expect,
-            chr1_cn,
-            chrx_cn,
-            chry_cn,
-        ) in (
-            ("formats/f-on-f.cns", True, False, 0, 0, None, 2, 2, None),
-            ("formats/f-on-m.cns", True, True, 0.585, 1, None, 3, 2, None),
-            ("formats/m-on-f.cns", False, False, 0, -1, 0, 2, 1, 1),
-            ("formats/m-on-m.cns", False, True, 0, 0, 0, 2, 1, 1),
-        ):
-            cns = cnvlib.read(fname)
-            chr1_idx = cns.chromosome == "chr1"
-            chrx_idx = cns.chromosome == "chrX"
-            chry_idx = cns.chromosome == "chrY"
-
-            def test_chrom_means(segments):
-                self.assertEqual(chr1_cn, segments["cn"][chr1_idx].mean())
-                self.assertAlmostEqual(
-                    chr1_expect, segments["log2"][chr1_idx].mean(), 0
-                )
-                self.assertEqual(chrx_cn, segments["cn"][chrx_idx].mean())
-                self.assertAlmostEqual(
-                    chrx_expect, segments["log2"][chrx_idx].mean(), 0
-                )
-                if not sample_is_f:
-                    self.assertEqual(chry_cn, segments["cn"][chry_idx].mean())
-                    self.assertAlmostEqual(
-                        chry_expect, segments["log2"][chry_idx].mean(), 0
-                    )
-
-            # Call threshold
-            cns_thresh = commands.do_call(
-                cns,
-                None,
-                "threshold",
-                is_haploid_x_reference=ref_is_m,
-                is_sample_female=sample_is_f,
-            )
-            test_chrom_means(cns_thresh)
-            # Call clonal pure
-            cns_clone = commands.do_call(
-                cns,
-                None,
-                "clonal",
-                is_haploid_x_reference=ref_is_m,
-                is_sample_female=sample_is_f,
-            )
-            test_chrom_means(cns_clone)
-            # Call clonal barely-mixed
-            cns_p99 = commands.do_call(
-                cns,
-                None,
-                "clonal",
-                purity=0.99,
-                is_haploid_x_reference=ref_is_m,
-                is_sample_female=sample_is_f,
-            )
-            test_chrom_means(cns_p99)
-
-    def test_call_various_abs_ref_exp_methods(self):
-        cnarr = cnvlib.read("formats/par-reference.grch38.cnn")
-
-        def _run(is_haploid_x_reference, is_sample_female, diploid_parx_genome=None):
-            ploidy = 2
-            purity = 0.8
-            abs_df = call.absolute_dataframe(
-                cnarr,
-                ploidy,
-                purity,
-                is_haploid_x_reference,
-                diploid_parx_genome,
-                is_sample_female,
-            )
-            abs_ref = call.absolute_reference(
-                cnarr, ploidy, diploid_parx_genome, is_haploid_x_reference
-            )
-            abs_exp = call.absolute_expect(
-                cnarr, ploidy, diploid_parx_genome, is_sample_female
-            )
-            abs_clonal = call.absolute_clonal(
-                cnarr,
-                ploidy,
-                purity,
-                is_haploid_x_reference,
-                diploid_parx_genome,
-                is_sample_female,
-            )
-            return abs_df, abs_ref, abs_exp, abs_clonal
-
-        def _assert_abs_df(iloc, abs_df, ref_copies, exp_copies):
-            self.assertTrue("reference" in abs_df.columns)
-            self.assertTrue("expect" in abs_df.columns)
-            r = abs_df.iloc[iloc]
-            self.assertEqual(r.reference, ref_copies)
-            self.assertEqual(r.expect, exp_copies)
-
-        def _assert_abs_copies(i, abs_values, copies):
-            self.assertEqual(abs_values[i], copies)
-
-        def _assert_abs_clonal(i, abs_clonal, value):
-            self.assertAlmostEqual(abs_clonal[i], value, 5)
-
-        def _assert_chr1(abs_df, abs_ref, abs_exp, abs_clonal):
-            i = 0
-            _assert_abs_df(i, abs_df, 2, 2)
-            _assert_abs_copies(i, abs_ref, 2)
-            _assert_abs_copies(i, abs_exp, 2)
-            _assert_abs_clonal(i, abs_clonal, 0.26708)
-
-        def _assert_chrx_par(
-            abs_df, abs_ref, abs_exp, abs_clonal, ref_copies, exp_copies, clonal_copies
-        ):
-            i = 13
-            _assert_abs_df(i, abs_df, ref_copies, exp_copies)
-            _assert_abs_copies(i, abs_ref, ref_copies)
-            _assert_abs_copies(i, abs_exp, exp_copies)
-            _assert_abs_clonal(i, abs_clonal, clonal_copies)
-
-        def _assert_chrx_non_par(
-            abs_df, abs_ref, abs_clonal, abs_exp, ref_copies, exp_copies, clonal_copies
-        ):
-            i = 21
-            _assert_abs_df(i, abs_df, ref_copies, exp_copies)
-            _assert_abs_copies(i, abs_ref, ref_copies)
-            _assert_abs_copies(i, abs_exp, exp_copies)
-            _assert_abs_clonal(i, abs_clonal, clonal_copies)
-
-        def _assert_chry_par(
-            abs_df, abs_ref, abs_exp, abs_clonal, ref_copies, exp_copies, clonal_copies
-        ):
-            i = 36
-            _assert_abs_df(i, abs_df, ref_copies, exp_copies)
-            _assert_abs_copies(i, abs_ref, ref_copies)
-            _assert_abs_copies(i, abs_exp, exp_copies)
-            _assert_abs_clonal(i, abs_clonal, clonal_copies)
-
-        def _assert_chry_non_par(
-            abs_df, abs_ref, abs_exp, abs_clonal, ref_copies, exp_copies, clonal_copies
-        ):
-            i = 40
-            _assert_abs_df(i, abs_df, ref_copies, exp_copies)
-            _assert_abs_copies(i, abs_ref, ref_copies)
-            _assert_abs_copies(i, abs_exp, exp_copies)
-            _assert_abs_clonal(i, abs_clonal, clonal_copies)
-
-        is_haploid_x_reference = True
-        is_female_sample = True
-        abs_df, abs_ref, abs_exp, abs_clonal = _run(
-            is_haploid_x_reference, is_female_sample
-        )
-        _assert_chr1(abs_df, abs_ref, abs_exp, abs_clonal)
-        _assert_chrx_par(abs_df, abs_ref, abs_exp, abs_clonal, 1, 2, 1.59225)
-        _assert_chrx_non_par(abs_df, abs_ref, abs_clonal, abs_exp, 1, 2, 1.34001)
-        _assert_chry_par(abs_df, abs_ref, abs_exp, abs_clonal, 1, 0, 2.04202)
-        _assert_chry_non_par(abs_df, abs_ref, abs_exp, abs_clonal, 1, 0, 0.70083)
-        abs_df, abs_ref, abs_exp, abs_clonal = _run(
-            is_haploid_x_reference, is_female_sample, "grch38"
-        )
-        _assert_chr1(abs_df, abs_ref, abs_exp, abs_clonal)
-        _assert_chrx_par(abs_df, abs_ref, abs_exp, abs_clonal, 2, 2, 3.68449)
-        _assert_chrx_non_par(abs_df, abs_ref, abs_clonal, abs_exp, 1, 2, 1.34001)
-        _assert_chry_par(abs_df, abs_ref, abs_exp, abs_clonal, 0, 0, 0.0)
-        _assert_chry_non_par(abs_df, abs_ref, abs_exp, abs_clonal, 1, 0, 0.70083)
-
-        is_haploid_x_reference = True
-        is_female_sample = False
-        abs_df, abs_ref, abs_exp, abs_clonal = _run(
-            is_haploid_x_reference, is_female_sample
-        )
-        _assert_chr1(abs_df, abs_ref, abs_exp, abs_clonal)
-        _assert_chrx_par(abs_df, abs_ref, abs_exp, abs_clonal, 1, 1, 1.84225)
-        _assert_chrx_non_par(abs_df, abs_ref, abs_clonal, abs_exp, 1, 1, 1.59001)
-        _assert_chry_par(abs_df, abs_ref, abs_exp, abs_clonal, 1, 1, 1.79202)
-        _assert_chry_non_par(abs_df, abs_ref, abs_exp, abs_clonal, 1, 1, 0.45083)
-        abs_df, abs_ref, abs_exp, abs_clonal = _run(
-            is_haploid_x_reference, is_female_sample, "grch38"
-        )
-        _assert_chr1(abs_df, abs_ref, abs_exp, abs_clonal)
-        _assert_chrx_par(abs_df, abs_ref, abs_exp, abs_clonal, 2, 2, 3.68449)
-        _assert_chrx_non_par(abs_df, abs_ref, abs_clonal, abs_exp, 1, 1, 1.59001)
-        _assert_chry_par(abs_df, abs_ref, abs_exp, abs_clonal, 0, 0, 0.0)
-        _assert_chry_non_par(abs_df, abs_ref, abs_exp, abs_clonal, 1, 1, 0.45083)
-
-        is_haploid_x_reference = False
-        is_female_sample = True
-        abs_df, abs_ref, abs_exp, abs_clonal = _run(
-            is_haploid_x_reference, is_female_sample
-        )
-        _assert_chr1(abs_df, abs_ref, abs_exp, abs_clonal)
-        _assert_chrx_par(abs_df, abs_ref, abs_exp, abs_clonal, 2, 2, 3.68449)
-        _assert_chrx_non_par(abs_df, abs_ref, abs_clonal, abs_exp, 2, 2, 3.18002)
-        _assert_chry_par(abs_df, abs_ref, abs_exp, abs_clonal, 1, 0, 2.04202)
-        _assert_chry_non_par(abs_df, abs_ref, abs_exp, abs_clonal, 1, 0, 0.70083)
-        abs_df, abs_ref, abs_exp, abs_clonal = _run(
-            is_haploid_x_reference, is_female_sample, "grch38"
-        )
-        _assert_chr1(abs_df, abs_ref, abs_exp, abs_clonal)
-        _assert_chrx_par(abs_df, abs_ref, abs_exp, abs_clonal, 2, 2, 3.684493)
-        _assert_chrx_non_par(abs_df, abs_ref, abs_clonal, abs_exp, 2, 2, 3.18002)
-        _assert_chry_par(abs_df, abs_ref, abs_exp, abs_clonal, 0, 0, 0.0)
-        _assert_chry_non_par(abs_df, abs_ref, abs_exp, abs_clonal, 1, 0, 0.70083)
-
-        is_haploid_x_reference = False
-        is_female_sample = False
-        abs_df, abs_ref, abs_exp, abs_clonal = _run(
-            is_haploid_x_reference, is_female_sample
-        )
-        _assert_chr1(abs_df, abs_ref, abs_exp, abs_clonal)
-        _assert_chrx_par(abs_df, abs_ref, abs_exp, abs_clonal, 2, 1, 3.93449)
-        _assert_chrx_non_par(abs_df, abs_ref, abs_clonal, abs_exp, 2, 1, 3.43002)
-        _assert_chry_par(abs_df, abs_ref, abs_exp, abs_clonal, 1, 1, 1.79202)
-        _assert_chry_non_par(abs_df, abs_ref, abs_exp, abs_clonal, 1, 1, 0.45083)
-        abs_df, abs_ref, abs_exp, abs_clonal = _run(
-            is_haploid_x_reference, is_female_sample, "grch38"
-        )
-        _assert_chr1(abs_df, abs_ref, abs_exp, abs_clonal)
-        _assert_chrx_par(abs_df, abs_ref, abs_exp, abs_clonal, 2, 2, 3.68449)
-        _assert_chrx_non_par(abs_df, abs_ref, abs_clonal, abs_exp, 2, 1, 3.43002)
-        _assert_chry_par(abs_df, abs_ref, abs_exp, abs_clonal, 0, 0, 0.0)
-        _assert_chry_non_par(abs_df, abs_ref, abs_exp, abs_clonal, 1, 1, 0.45083)
-
-    @staticmethod
-    def _deep_deletion_cns():
-        """Segments including deep deletions: a true homozygous deletion
-        (log2=-3) and a zero-coverage sentinel segment (log2=NULL_LOG2_COVERAGE).
-        """
-        log2 = [0.0, -3.0, params.NULL_LOG2_COVERAGE, 0.585]
-        n = len(log2)
-        return cnary.CopyNumArray(
-            pd.DataFrame(
-                {
-                    "chromosome": ["chr1"] * n,
-                    "start": np.arange(0, n * 1000, 1000),
-                    "end": np.arange(1000, n * 1000 + 1000, 1000),
-                    "gene": ["-"] * n,
-                    "log2": log2,
-                    "probes": [10] * n,
-                    "weight": [1.0] * n,
-                }
-            )
-        )
-
-    def test_call_clonal_no_negative_cn(self):
-        """Clonal calling with impure samples never emits negative copy number.
-
-        Regression for #503/#516: with purity < 1, the purity-rescale formula
-        n = (r*2^log2 - x*(1-p)) / p extrapolates to negative absolute copies
-        for deeply deleted (or zero-coverage sentinel) segments. Absolute copy
-        number is physically >= 0, so it must be floored at 0.
-        """
-        cns = self._deep_deletion_cns()
-        for tumor_purity in (0.3, 0.5, 0.65, 0.9):
-            with self.subTest(purity=tumor_purity):
-                called = commands.do_call(
-                    cns,
-                    None,
-                    "clonal",
-                    purity=tumor_purity,
-                    is_haploid_x_reference=True,
-                    is_sample_female=True,
-                )
-                self.assertTrue(
-                    (called["cn"] >= 0).all(),
-                    f"negative cn at purity={tumor_purity}: {called['cn'].tolist()}",
-                )
-                # Deep-deletion and sentinel segments should call CN 0
-                self.assertEqual(called["cn"].iloc[1], 0)
-                self.assertEqual(called["cn"].iloc[2], 0)
-
-    def test_log2_ratio_to_absolute_floored_at_zero(self):
-        """_log2_ratio_to_absolute never returns a negative absolute (#503)."""
-        # Impure path: deep deletion below the (1-p) contamination floor
-        self.assertEqual(call._log2_ratio_to_absolute(-20.0, 2, 2, purity=0.5), 0.0)
-        self.assertEqual(call._log2_ratio_to_absolute(-3.0, 2, 2, purity=0.5), 0.0)
-        # A real single-copy gain is unaffected
-        self.assertAlmostEqual(
-            call._log2_ratio_to_absolute(0.585, 2, 2, purity=1.0), 3.0, places=2
-        )
-        # A NaN log2 (malformed input) propagates rather than being silently
-        # floored to 0 -- a false homozygous-deletion call would be worse.
-        self.assertTrue(
-            np.isnan(call._log2_ratio_to_absolute(np.nan, 2, 2, purity=0.5))
-        )
-
-
-class LoadHetSnpsTests(unittest.TestCase):
-    """Tests for cmdutil.load_het_snps and the _warn_if_baf_input_suspicious helper."""
-
-    def test_helper_warns_on_empty(self):
-        """Empty heterozygous result emits a clear warning."""
-        with self.assertLogs(level="WARNING") as ctx:
-            cmdutil._warn_if_baf_input_suspicious(None)
-        self.assertTrue(
-            any("No heterozygous variants" in msg for msg in ctx.output),
-            ctx.output,
-        )
-
-    def test_helper_warns_on_skewed_distribution(self):
-        """Median alt_freq far from 0.5 emits a warning."""
-        # 100 variants all near 1.0 (e.g. mistakenly homozygous-alt or somatic)
-        alt_freqs = pd.Series([0.95] * 100)
-        with self.assertLogs(level="WARNING") as ctx:
-            cmdutil._warn_if_baf_input_suspicious(alt_freqs)
-        self.assertTrue(
-            any("Median allele frequency" in msg for msg in ctx.output),
-            ctx.output,
-        )
-
-    def test_helper_silent_on_balanced_distribution(self):
-        """Median alt_freq near 0.5 emits no warning."""
-        rng = np.random.default_rng(0)
-        # 100 het SNPs with alt_freq centered on 0.5 (binomial-ish noise)
-        alt_freqs = pd.Series(rng.normal(0.5, 0.05, 100).clip(0, 1))
-        with self.assertNoLogs(level="WARNING"):
-            cmdutil._warn_if_baf_input_suspicious(alt_freqs)
-
-    def test_helper_silent_below_min_count(self):
-        """Distribution check is skipped for small variant sets to avoid noise."""
-        # Skewed but only 10 variants -- no warning (could be a small panel)
-        alt_freqs = pd.Series([0.95] * 10)
-        with self.assertNoLogs(level="WARNING"):
-            cmdutil._warn_if_baf_input_suspicious(alt_freqs)
-
-    def test_load_het_snps_warns_on_empty_vcf(self):
-        """Loading a VCF with no records triggers the empty-result warning."""
-        with self.assertLogs(level="WARNING") as ctx:
-            varr = commands.load_het_snps("formats/blank.vcf", None, None, 1, None)
-        self.assertEqual(len(varr), 0)
-        self.assertTrue(
-            any("No heterozygous variants" in msg for msg in ctx.output),
-            ctx.output,
-        )
-
-    def test_load_het_snps_silent_on_real_vcf(self):
-        """A legitimate test VCF with paired sample IDs doesn't trigger warnings."""
-        with self.assertNoLogs(level="WARNING"):
-            varr = commands.load_het_snps(
-                "formats/na12878_na12882_mix.vcf", "NA12878", "NA12882", 15, None
-            )
-        self.assertGreater(len(varr), 50)
-
-    def test_load_het_snps_warns_on_missing_sample_ids(self):
-        """Without -i/-n, the wrong sample is used; the distribution warning fires."""
-        with self.assertLogs(level="WARNING") as ctx:
-            commands.load_het_snps(
-                "formats/na12878_na12882_mix.vcf", None, None, 15, None
-            )
-        self.assertTrue(
-            any("Median allele frequency" in msg for msg in ctx.output),
-            ctx.output,
-        )
-
-
-class AnalysisTests(unittest.TestCase):
-    """Tests for analysis commands: bintest, breaks, genemetrics, metrics, etc."""
-
-    def test_bintest(self):
-        """The 'bintest' command."""
-        cnarr = cnvlib.read("formats/amplicon.cnr")
-        segarr = cnvlib.read("formats/amplicon.cns")
-        # Simple
-        rows = commands.do_bintest(cnarr, alpha=0.05)
-        self.assertGreater(len(rows), 0)
-        self.assertLess(len(rows), len(cnarr))
-        # Versus segments
-        rows = commands.do_bintest(cnarr, segarr, target_only=True)
-        self.assertGreaterEqual(len(rows), len(segarr))
-        self.assertLess(len(rows), len(cnarr))
-
-    def test_breaks(self):
-        """The 'breaks' command."""
-        probes = cnvlib.read("formats/amplicon.cnr")
-        segs = cnvlib.read("formats/amplicon.cns")
-        rows = commands.do_breaks(probes, segs, 4)
-        self.assertGreater(len(rows), 0)
-
-    def test_genemetrics(self):
-        """The 'genemetrics' command."""
-        probes = cnvlib.read("formats/amplicon.cnr")
-        rows = commands.do_genemetrics(probes, is_haploid_x_reference=True)
-        self.assertGreater(len(rows), 0)
-        segs = cnvlib.read("formats/amplicon.cns")
-        rows = commands.do_genemetrics(
-            probes, segs, 0.3, 4, is_haploid_x_reference=True
-        )
-        self.assertGreater(len(rows), 0)
-
-    @pytest.mark.slow
-    def test_genemetrics_with_stats(self):
-        """The 'genemetrics' command with statistics options."""
-        probes = cnvlib.read("formats/amplicon.cnr")
-        segs = cnvlib.read("formats/amplicon.cns")
-
-        # Test location statistics
-        result = commands.do_genemetrics(
-            probes,
-            segs,
-            0.3,
-            4,
-            is_haploid_x_reference=True,
-            location_stats=["mean", "median", "mode", "p_ttest"],
-        )
-        self.assertGreater(len(result), 0)
-        self.assertIn("mean", result.columns)
-        self.assertIn("median", result.columns)
-        self.assertIn("mode", result.columns)
-        self.assertIn("p_ttest", result.columns)
-        # P-values should be between 0 and 1 (excluding NaN)
-        valid_p = result["p_ttest"].dropna()
-        self.assertGreater(len(valid_p), 0)
-        self.assertTrue((valid_p >= 0).all())
-        self.assertTrue((valid_p <= 1).all())
-
-        # Test spread statistics
-        result = commands.do_genemetrics(
-            probes,
-            segs,
-            0.3,
-            4,
-            is_haploid_x_reference=True,
-            spread_stats=["stdev", "sem", "mad", "mse", "iqr", "bivar"],
-        )
-        self.assertGreater(len(result), 0)
-        for stat in ["stdev", "sem", "mad", "mse", "iqr", "bivar"]:
-            self.assertIn(stat, result.columns)
-            # Spread statistics should be non-negative (excluding NaN)
-            valid_stat = result[stat].dropna()
-            if len(valid_stat) > 0:
-                self.assertTrue((valid_stat >= 0).all())
-
-        # Test interval statistics with adaptive smoothed bootstrap
-        result = commands.do_genemetrics(
-            probes,
-            segs,
-            0.3,
-            4,
-            is_haploid_x_reference=True,
-            location_stats=["mean", "median"],
-            interval_stats=["ci", "pi"],
-            alpha=0.05,
-            bootstraps=50,
-            smoothed=10,  # Default threshold
-        )
-        self.assertGreater(len(result), 0)
-        self.assertIn("ci_lo", result.columns)
-        self.assertIn("ci_hi", result.columns)
-        self.assertIn("pi_lo", result.columns)
-        self.assertIn("pi_hi", result.columns)
-
-        # Confidence intervals should be ordered: ci_lo <= mean <= ci_hi
-        # Filter out rows with NaN
-        valid_rows = result.dropna(subset=["ci_lo", "ci_hi", "mean"])
-        self.assertGreater(len(valid_rows), 0)
-        ci_contains_mean = (valid_rows["ci_lo"] <= valid_rows["mean"]) & (
-            valid_rows["mean"] <= valid_rows["ci_hi"]
-        )
-        # Most should contain the mean (allow some misses due to random sampling)
-        self.assertGreater(ci_contains_mean.sum(), len(valid_rows) * 0.9)
-
-        # Prediction intervals should contain the mean
-        valid_rows = result.dropna(subset=["pi_lo", "pi_hi", "mean"])
-        self.assertGreater(len(valid_rows), 0)
-        pi_contains_mean = (valid_rows["pi_lo"] <= valid_rows["mean"]) & (
-            valid_rows["mean"] <= valid_rows["pi_hi"]
-        )
-        self.assertGreater(pi_contains_mean.sum(), len(valid_rows) * 0.9)
-
-        # Intervals should be properly ordered (lo <= hi)
-        valid_rows = result.dropna(subset=["pi_lo", "pi_hi", "ci_lo", "ci_hi"])
-        self.assertGreater(len(valid_rows), 0)
-        self.assertTrue((valid_rows["ci_lo"] <= valid_rows["ci_hi"]).all())
-        self.assertTrue((valid_rows["pi_lo"] <= valid_rows["pi_hi"]).all())
-        # PI is usually wider than CI, but may not always be for very small
-        # segments with smoothed bootstrap, so just check most cases
-        pi_width = valid_rows["pi_hi"] - valid_rows["pi_lo"]
-        ci_width = valid_rows["ci_hi"] - valid_rows["ci_lo"]
-        self.assertGreater((pi_width >= ci_width).sum(), len(valid_rows) * 0.8)
-
-        # Test without segments (gene-level analysis)
-        result = commands.do_genemetrics(
-            probes,
-            None,
-            0.3,
-            4,
-            is_haploid_x_reference=True,
-            location_stats=["mean", "median"],
-            spread_stats=["stdev"],
-            interval_stats=["pi"],
-        )
-        self.assertGreater(len(result), 0)
-        self.assertIn("mean", result.columns)
-        self.assertIn("stdev", result.columns)
-        self.assertIn("pi_lo", result.columns)
-        self.assertIn("pi_hi", result.columns)
-
-        # Test different smoothed bootstrap thresholds
-        # Small threshold (0) - should use BCa for all
-        result_bca = commands.do_genemetrics(
-            probes,
-            segs,
-            0.3,
-            4,
-            is_haploid_x_reference=True,
-            location_stats=["mean"],
-            interval_stats=["ci"],
-            bootstraps=50,
-            smoothed=0,  # Never smooth, always BCa
-        )
-        self.assertGreater(len(result_bca), 0)
-
-        # Large threshold (1000) - should use smoothed bootstrap for all
-        result_smooth = commands.do_genemetrics(
-            probes,
-            segs,
-            0.3,
-            4,
-            is_haploid_x_reference=True,
-            location_stats=["mean"],
-            interval_stats=["ci"],
-            bootstraps=50,
-            smoothed=1000,  # Always smooth
-        )
-        self.assertGreater(len(result_smooth), 0)
-
-        # CIs should be valid (lo <= hi)
-        valid_bca = result_bca.dropna(subset=["ci_lo", "ci_hi"])
-        valid_smooth = result_smooth.dropna(subset=["ci_lo", "ci_hi"])
-        self.assertGreater(len(valid_bca), 0)
-        self.assertGreater(len(valid_smooth), 0)
-        self.assertTrue((valid_bca["ci_lo"] <= valid_bca["ci_hi"]).all())
-        self.assertTrue((valid_smooth["ci_lo"] <= valid_smooth["ci_hi"]).all())
-
-    def test_import_theta(self):
-        """The 'import-theta' command."""
-        cns = cnvlib.read("formats/nv3.cns")
-        theta_fname = "formats/nv3.n3.results"
-        for new_cns in commands.do_import_theta(cns, theta_fname):
-            self.assertTrue(0 < len(new_cns) <= len(cns))
-
-    def test_metrics(self):
-        """The 'metrics' command."""
-        cnarr = cnvlib.read("formats/amplicon.cnr")
-        segments = cnvlib.read("formats/amplicon.cns")
-        result = metrics.do_metrics(cnarr, segments, skip_low=True)
-        self.assertEqual(result.shape, (1, 6))
-        values = result.loc[0, result.columns[1:]]
-        for val in values:
-            self.assertGreater(val, 0)
-
-    def test_segmetrics(self):
-        """The 'segmetrics' command."""
-        cnarr = cnvlib.read("formats/amplicon.cnr")
-        segarr = cnvlib.read("formats/amplicon.cns")
-        sm = segmetrics.do_segmetrics(
-            cnarr,
-            segarr,
-            location_stats=["mean", "median", "mode", "p_ttest"],
-            spread_stats=["stdev", "sem", "iqr"],
-            interval_stats=["pi", "ci"],
-            bootstraps=50,
-            smoothed=True,
-        )
-        # Restrict to segments with enough supporting probes for sane stats
-        sm = sm[sm["probes"] > 3]
-        self.assertTrue((sm["pi_lo"] < sm["median"]).all())
-        self.assertTrue((sm["pi_hi"] > sm["median"]).all())
-        self.assertTrue((sm["ci_lo"] < sm["mean"]).all())
-        self.assertTrue((sm["ci_hi"] > sm["mean"]).all())
-
-    @pytest.mark.slow
-    def test_purity(self):
-        """The 'purity' command."""
-        segments = cnvlib.read("formats/tr95t.cns")
-
-        # Basic call without VCF
-        result = commands.do_purity(segments)
-        self.assertIn("purity", result.columns)
-        self.assertIn("ploidy", result.columns)
-        self.assertIn("score", result.columns)
-        self.assertGreater(len(result), 0)
-        self.assertGreater(result["purity"].iloc[0], 0)
-        self.assertLessEqual(result["purity"].iloc[0], 1.0)
-        self.assertGreaterEqual(result["ploidy"].iloc[0], 1.5)
-        self.assertLessEqual(result["ploidy"].iloc[0], 5.0)
-        self.assertTrue((result["score"] > 0).all())
-
-        # With VCF
-        varr = tabio.read(
-            "formats/na12878_na12882_mix.vcf",
-            "vcf",
-            skip_somatic=True,
-        ).heterozygous()
-        result_baf = commands.do_purity(segments, varr)
-        self.assertEqual(list(result_baf.columns), ["purity", "ploidy", "score"])
-        self.assertGreater(len(result_baf), 0)
-
-        # Custom grid (coarser steps for speed)
-        result_coarse = commands.do_purity(
-            segments,
-            min_purity=0.2,
-            max_purity=0.8,
-            purity_step=0.1,
-            min_ploidy=2.0,
-            max_ploidy=4.0,
-            ploidy_step=0.5,
-        )
-        self.assertGreater(len(result_coarse), 0)
-        self.assertGreaterEqual(result_coarse["purity"].iloc[0], 0.2)
-        self.assertLessEqual(result_coarse["purity"].iloc[0], 0.8)
-
-    def test_purity_file(self):
-        """The 'purity' command: read_purity_tsv round-trip."""
-        with tempfile.NamedTemporaryFile(mode="w", suffix=".tsv", delete=False) as f:
-            f.write("purity\tploidy\tscore\n")
-            f.write("0.65\t2.1\t42.5\n")
-            f.write("0.70\t2.0\t41.0\n")
-            tmp_path = f.name
-        try:
-            pur, plo = purity.read_purity_tsv(tmp_path)
-            self.assertAlmostEqual(pur, 0.65)
-            self.assertAlmostEqual(plo, 2.1)
-        finally:
-            os.unlink(tmp_path)
-
-
-class PlotTests(unittest.TestCase):
-    """Smoke tests for plotting commands."""
-
-    def test_scatter(self):
-        """The 'scatter' command."""
-        cnarr = cnvlib.read("formats/amplicon.cnr")
-        segarr = cnvlib.read("formats/amplicon.cns")
-        fig = scatter.do_scatter(cnarr, segarr)
-        self.assertIsNotNone(fig)
-        # With a gene zoom
-        fig = scatter.do_scatter(cnarr, segarr, show_gene="BRAF")
-        self.assertIsNotNone(fig)
-
-    def test_scatter_genome_y_floor(self):
-        """Genome-wide autoscale floors y_min so a single deep homozygous
-        deletion can't compress the whole plot (gh#385)."""
-        from matplotlib import pyplot
-
-        probes = cnary.CopyNumArray.from_rows(
-            [
-                ["chr1", 100, 200, "A", 0.0],
-                ["chr1", 200, 300, "A", -0.1],
-                ["chr1", 300, 400, "A", -12.0],
-                ["chr2", 100, 200, "B", 0.05],
-                ["chr2", 200, 300, "B", -0.2],
-            ],
-            columns=["chromosome", "start", "end", "gene", "log2"],
-        )
-        segments = cnary.CopyNumArray.from_rows(
-            [
-                ["chr1", 100, 400, "A", -12.0],
-                ["chr2", 100, 300, "B", 0.0],
-            ],
-            columns=["chromosome", "start", "end", "gene", "log2"],
-        )
-        segments.data["probes"] = [3, 2]
-        _fig, ax = pyplot.subplots()
-        scatter.cnv_on_genome(ax, probes, segments)
-        y_lo, _y_hi = ax.get_ylim()
-        pyplot.close(_fig)
-        self.assertGreaterEqual(
-            y_lo,
-            scatter.AUTO_Y_MIN_FLOOR,
-            "Deep deletion must not pull the genome-wide y-axis below the floor",
-        )
-        # An explicit --y-min still overrides the floor
-        _fig, ax = pyplot.subplots()
-        scatter.cnv_on_genome(ax, probes, segments, y_min=-15.0, y_max=2.0)
-        y_lo, _y_hi = ax.get_ylim()
-        pyplot.close(_fig)
-        self.assertAlmostEqual(y_lo, -15.0)
-
-    def test_heatmap(self):
-        """The 'heatmap' command."""
-        cnarrs = [cnvlib.read("formats/amplicon.cnr")]
-        ax = heatmap.do_heatmap(cnarrs)
-        self.assertIsNotNone(ax)
-        # With desaturation
-        ax = heatmap.do_heatmap(cnarrs, do_desaturate=True)
-        self.assertIsNotNone(ax)
-
-
-class ExportTests(unittest.TestCase):
-    """Tests for export commands."""
-
-    def test_export_bed_vcf(self):
-        """The 'export' command for formats with absolute copy number."""
-        for fname, ploidy, is_f in [
-            ("tr95t.cns", 2, True),
-            ("cl_seq.cns", 6, True),
-            ("amplicon.cns", 2, False),
-        ]:
-            with self.subTest(fname=fname, ploidy=ploidy):
-                cns = cnvlib.read("formats/" + fname)
-                # BED
-                for show in ("ploidy", "variant", "all"):
-                    tbl_bed = export.export_bed(
-                        cns, ploidy, True, None, is_f, cns.sample_id, show
-                    )
-                    if show == "all":
-                        self.assertEqual(len(tbl_bed), len(cns), f"{fname} {ploidy}")
-                    else:
-                        self.assertLess(len(tbl_bed), len(cns))
-                # VCF
-                _vheader, vcf_body = export.export_vcf(cns, ploidy, True, None, is_f)
-                self.assertTrue(0 < len(vcf_body.splitlines()) < len(cns))
-
-    def test_export_cdt_jtv(self):
-        """The 'export' command for CDT and Java TreeView formats."""
-        fnames = ["formats/p2-20_1.cnr", "formats/p2-20_2.cnr"]
-        sample_ids = list(map(core.fbase, fnames))
-        nrows = linecount(fnames[0]) - 1
-        for fmt_key, header2 in (("cdt", 2), ("jtv", 0)):
-            table = export.merge_samples(fnames)
-            formatter = export.EXPORT_FORMATS[fmt_key]
-            _oh, outrows = formatter(sample_ids, table)
-            self.assertEqual(len(list(outrows)), nrows + header2)
-
-    def test_export_nexus(self):
-        """The 'export nexus-basic' and 'nexus-ogt' commands."""
-        cnr = cnvlib.read("formats/amplicon.cnr")
-        table_nb = export.export_nexus_basic(cnr)
-        self.assertEqual(len(table_nb), len(cnr))
-        varr = commands.load_het_snps(
-            "formats/na12878_na12882_mix.vcf", None, None, 15, None
-        )
-        table_ogt = export.export_nexus_ogt(cnr, varr, 0.05)
-        self.assertEqual(len(table_ogt), len(cnr))
-
-    def test_export_seg(self):
-        """The 'export seg' command."""
-        seg_rows = export.export_seg(["formats/tr95t.cns"])
-        self.assertGreater(len(seg_rows), 0)
-        seg2_rows = export.export_seg(["formats/tr95t.cns", "formats/cl_seq.cns"])
-        self.assertGreater(len(seg2_rows), len(seg_rows))
-
-    def test_export_theta(self):
-        """The 'export theta' command."""
-        segarr = cnvlib.read("formats/tr95t.cns")
-        len_seg_auto = len(segarr.autosomes())
-        table_theta = export.export_theta(segarr, None)
-        self.assertEqual(len(table_theta), len_seg_auto)
-        ref = cnvlib.read("formats/reference-tr.cnn")
-        table_theta = export.export_theta(segarr, ref)
-        self.assertEqual(len(table_theta), len_seg_auto)
-        varr = commands.load_het_snps(
-            "formats/na12878_na12882_mix.vcf", "NA12882", "NA12878", 15, None
-        )
-        tumor_snps, normal_snps = export.export_theta_snps(varr)
-        self.assertLess(len(tumor_snps), len(varr))
-        self.assertGreater(len(tumor_snps), 0)
-        self.assertLess(len(normal_snps), len(varr))
-        self.assertGreater(len(normal_snps), 0)
-
-
-class BatchTests(unittest.TestCase):
-    """Tests for the batch command and end-to-end pipeline."""
-
-    def test_batch(self):
-        """The 'batch' command."""
-        target_bed = "formats/my-targets.bed"
-        fasta = "formats/chrM-Y-trunc.hg19.fa"
-        bam = "formats/na12878-chrM-Y-trunc.bam"
-        annot = "formats/my-refflat.bed"
-        # Build a single-sample WGS reference
-        ref_fname, tgt_bed_fname, _ = batch.batch_make_reference(
-            [bam],
-            None,
-            None,
-            True,
-            None,
-            fasta,
-            annot,
-            True,
-            500,
-            None,
-            None,
-            None,
-            None,
-            "build",
-            1,
-            False,
-            0,
-            "wgs",
-            False,
-        )
-        self.assertEqual(ref_fname, "build/reference.cnn")
-        refarr = cnvlib.read(ref_fname, "bed")
-        tgt_regions = tabio.read(tgt_bed_fname, "bed")
-        self.assertEqual(len(refarr), len(tgt_regions))
-        # Build a single-sample hybrid-capture reference
-        ref_fname, tgt_bed_fname, anti_bed_fname = batch.batch_make_reference(
-            [bam],
-            target_bed,
-            None,
-            True,
-            None,
-            fasta,
-            None,
-            True,
-            10,
-            None,
-            1000,
-            100,
-            None,
-            "build",
-            1,
-            False,
-            0,
-            "hybrid",
-            False,
-        )
-        self.assertEqual(ref_fname, "build/reference.cnn")
-        refarr = cnvlib.read(ref_fname, "bed")
-        tgt_regions = tabio.read(tgt_bed_fname, "bed")
-        anti_regions = tabio.read(anti_bed_fname, "bed")
-        self.assertEqual(len(refarr), len(tgt_regions) + len(anti_regions))
-        # Run the same sample
-        batch.batch_run_sample(
-            bam,
-            tgt_bed_fname,
-            anti_bed_fname,
-            ref_fname,
-            "build",
-            True,
-            None,
-            True,
-            True,
-            "Rscript",
-            False,
-            0,
-            False,
-            "hybrid",
-            "hmm",
-            1,
-            False,
-        )
-        cns = cnvlib.read("build/na12878-chrM-Y-trunc.cns")
-        self.assertGreater(len(cns), 0)
-
-    def test_batch_hybrid_autobin(self):
-        """Hybrid batch with autobin-determined bin sizes (issue #302)."""
-        target_bed = "formats/my-targets.bed"
-        fasta = "formats/chrM-Y-trunc.hg19.fa"
-        bam = "formats/na12878-chrM-Y-trunc.bam"
-        # target_avg_size=0 triggers autobin for hybrid mode
-        ref_fname, tgt_bed_fname, anti_bed_fname = batch.batch_make_reference(
-            [bam],
-            target_bed,
-            None,
-            True,
-            None,
-            fasta,
-            None,
-            True,
-            0,
-            None,
-            None,
-            None,
-            None,
-            "build",
-            1,
-            False,
-            0,
-            "hybrid",
-            False,
-        )
-        self.assertEqual(ref_fname, "build/reference.cnn")
-        refarr = cnvlib.read(ref_fname, "bed")
-        tgt_regions = tabio.read(tgt_bed_fname, "bed")
-        anti_regions = tabio.read(anti_bed_fname, "bed")
-        self.assertGreater(len(tgt_regions), 0)
-        self.assertEqual(len(refarr), len(tgt_regions) + len(anti_regions))
-
-    def test_batch_futures_exception_handling(self):
-        """Test that batch command properly waits for and reports exceptions from parallel tasks."""
-
-        # Test that SerialPool propagates exceptions correctly
-        def failing_task():
-            raise ValueError("Simulated processing error")
-
-        pool = parallel.SerialPool()
-        future = pool.submit(failing_task)
-
-        # The future should capture the exception
-        with self.assertRaises(ValueError) as ctx:
-            future.result()
-        self.assertIn("Simulated processing error", str(ctx.exception))
-
-        # Test that successful tasks work correctly
-        def success_task(x):
-            return x * 2
-
-        future2 = pool.submit(success_task, 21)
-        result = future2.result()
-        self.assertEqual(result, 42)
-
-    def test_batch_run_sample_with_duplicate_coordinates(self):
-        """Integration test: batch_run_sample with duplicate coordinates in coverage.
-
-        This tests that when coverage data contains duplicate coordinates,
-        the error is properly detected and reported with sample context.
-        """
-        # Create a temporary directory for output
-        with tempfile.TemporaryDirectory() as tmpdir:
-            # Create mock coverage data with duplicate coordinates
-            # This simulates the issue from GitHub issue #971
-            target_data = pd.DataFrame(
-                {
-                    "chromosome": ["chr1"] * 4,
-                    "start": [100, 200, 200, 300],  # Duplicate start at position 200
-                    "end": [150, 250, 250, 350],  # Duplicate end at position 250
-                    "gene": ["GeneA", "GeneB", "GeneB", "GeneC"],
-                    "log2": [0.1, 0.2, 0.2, 0.3],
-                    "depth": [100.0, 200.0, 200.0, 150.0],
-                }
-            )
-            target_cnarr = cnary.CopyNumArray(target_data, {"sample_id": "test_sample"})
-
-            antitarget_data = pd.DataFrame(
-                {
-                    "chromosome": ["chr1"] * 2,
-                    "start": [50, 400],
-                    "end": [90, 450],
-                    "gene": ["Antitarget", "Antitarget"],
-                    "log2": [0.0, 0.0],
-                    "depth": [50.0, 50.0],
-                }
-            )
-            antitarget_cnarr = cnary.CopyNumArray(
-                antitarget_data, {"sample_id": "test_sample"}
-            )
-
-            # Create a simple reference without duplicates
-            ref_data = pd.DataFrame(
-                {
-                    "chromosome": ["chr1"] * 5,
-                    "start": [50, 100, 200, 300, 400],
-                    "end": [90, 150, 250, 350, 450],
-                    "gene": ["Antitarget", "GeneA", "GeneB", "GeneC", "Antitarget"],
-                    "log2": [0.0, 0.0, 0.0, 0.0, 0.0],
-                }
-            )
-            ref_cnarr = cnary.CopyNumArray(ref_data, {"sample_id": "reference"})
-
-            # Test that do_fix detects duplicate coordinates and raises ValueError
-            with self.assertRaises(ValueError) as ctx:
-                fix.do_fix(target_cnarr, antitarget_cnarr, ref_cnarr)
-
-            # The error message should mention duplicates
-            self.assertIn("Duplicated genomic coordinates", str(ctx.exception))
-
-    @pytest.mark.slow
-    def test_diploid_parx_genome(self):
-        genome_build = "grch37"
-        male_target_fnames = [
-            "formats/male01.targetcoverage.cnn",
-            "formats/male02.targetcoverage.cnn",
-        ]
-        male_antitarget_fnames = [
-            "formats/male01.antitargetcoverage.cnn",
-            "formats/male02.antitargetcoverage.cnn",
-        ]
-        female_target_fnames = [
-            "formats/female01.targetcoverage.cnn",
-            "formats/female02.targetcoverage.cnn",
-        ]
-        female_antitarget_fnames = [
-            "formats/female01.antitargetcoverage.cnn",
-            "formats/female02.antitargetcoverage.cnn",
-        ]
-
-        # Baseline assumption: PAR1/2 is highly coveraged in the male sample.
-        target_cnn = cnvlib.read(male_target_fnames[0])
-        antitarget_cnn = cnvlib.read(male_antitarget_fnames[0])
-        target_cnn_non_parx = target_cnn[target_cnn.chr_x_filter(genome_build)]
-        antitarget_cnn_non_parx = antitarget_cnn[
-            antitarget_cnn.chr_x_filter(genome_build)
-        ]
-        target_cnn_parx = target_cnn[target_cnn.parx_filter(genome_build)]
-        antitarget_cnn_parx = antitarget_cnn[antitarget_cnn.parx_filter(genome_build)]
-        target_log2_mean_parx = target_cnn_parx["log2"].mean()
-        target_log2_mean_non_parx = target_cnn_non_parx["log2"].mean()
-        antitarget_log2_mean_parx = antitarget_cnn_parx["log2"].mean()
-        antitarget_log2_mean_non_parx = antitarget_cnn_non_parx["log2"].mean()
-        self.assertTrue(
-            target_log2_mean_parx - 0.9 > target_log2_mean_non_parx,
-            "PAR1/2 have nearly doubled coverage.",
-        )
-        self.assertTrue(
-            antitarget_log2_mean_parx - 0.9 > antitarget_log2_mean_non_parx,
-            "PAR1/2 have nearly doubled coverage.",
-        )
-
-        #### MIXED POOL ####
-        target_fnames = male_target_fnames + female_target_fnames
-        antitarget_fnames = male_antitarget_fnames + female_antitarget_fnames
-        # Only process male02 (index 1) and female01 (index 2) — the samples
-        # actually checked below — to avoid redundant CBS segmentation.
-        _ref_probes1, _cnrs1, _cnss1, clls1, _sex_df1 = run_samples(
-            target_fnames, antitarget_fnames, None, sample_indices=[1, 2]
-        )
-        _ref_probes2, _cnrs2, _cnss2, clls2, _sex_df2 = run_samples(
-            target_fnames, antitarget_fnames, genome_build, sample_indices=[1, 2]
-        )
-
-        # "dpxg" = DiploidParXGenome
-        male_call, male_call_dpxg = clls1[0], clls2[0]
-        female_call, female_call_dpxg = clls1[1], clls2[1]
-        male_call__x, male_call_dpxg__x = (
-            male_call[male_call.chr_x_filter()],
-            male_call_dpxg[male_call_dpxg.chromosome == "X"],
-        )
-        female_call__x, female_call_dpxg__x = (
-            female_call[female_call.chromosome == "X"],
-            female_call_dpxg[female_call_dpxg.chromosome == "X"],
-        )
-
-        # The cn=0 segment for the male sample is derived from a true loss on XAGE1B.
-        self.assertEqual(
-            male_call__x["cn"].to_list(),
-            [1, 1, 0, 1, 1],
-            "Non-Dpxg male has cn=1 including PAR1/2.",
-        )
-        self.assertEqual(
-            male_call_dpxg__x["cn"].to_list(),
-            [2, 1, 0, 1, 1, 2],
-            "Dpxg male has cn=2 for PAR1/2 and cn=1 otherwise.",
-        )
-        self.assertEqual(
-            female_call__x["cn"].to_list(),
-            [1, 2, 2, 1],
-            "Non-Dpxg female is biased towards loss in PAR1/2.",
-        )
-        self.assertEqual(
-            female_call_dpxg__x["cn"].to_list(),
-            [2, 2],
-            "Dpxg female has cn=2 everywhere including PAR1/2.",
-        )
-
-        #### MALE-ONLY POOL ####
-        target_fnames = male_target_fnames
-        antitarget_fnames = male_antitarget_fnames
-        # Only process male02 (index 1) — the sample actually checked below.
-        _ref_probes1, _cnrs1, _cnss1, clls1, _sex_df1 = run_samples(
-            target_fnames, antitarget_fnames, None, sample_indices=[1]
-        )
-        _ref_probes2, _cnrs2, _cnss2, clls2, _sex_df2 = run_samples(
-            target_fnames, antitarget_fnames, genome_build, sample_indices=[1]
-        )
-
-        male_call, male_call_dpxg = clls1[0], clls2[0]
-        male_call__x, male_call_dpxg__x = (
-            male_call[male_call.chr_x_filter()],
-            male_call_dpxg[male_call_dpxg.chromosome == "X"],
-        )
-        self.assertEqual(
-            male_call__x["cn"].to_list(),
-            [1, 1],
-            "Non-Dpxg male has cn=1 including PAR1/2.",
-        )
-        self.assertEqual(
-            male_call_dpxg__x["cn"].to_list(),
-            [2, 1, 1, 2],
-            "Dpxg male has cn=2 for PAR1/2 and cn=1 otherwise.",
-        )
-
-
-# == helpers ==
-
-
-def run_samples(
-    target_fnames, antitarget_fnames, diploid_parx_genome, sample_indices=None
-):
-    ref_probes = commands.do_reference(
-        target_fnames, antitarget_fnames, diploid_parx_genome=diploid_parx_genome
-    )
-    if sample_indices is None:
-        sample_indices = range(len(target_fnames))
-    cnrs, cnss, clls = [], [], []
-    for i in sample_indices:
-        cnr, cns, cll = run_sample(
-            target_fnames[i], antitarget_fnames[i], ref_probes, diploid_parx_genome
-        )
-        cnrs.append(cnr)
-        cnss.append(cns)
-        clls.append(cll)
-
-    sex_df = commands.do_sex(
-        cnrs, is_haploid_x_reference=False, diploid_parx_genome=diploid_parx_genome
-    )
-    return ref_probes, cnrs, cnss, clls, sex_df
-
-
-def run_sample(target_fname, antitarget_fname, ref_probes, diploid_parx_genome):
-    tgt_raw = cnvlib.read(target_fname)
-    anti_raw = cnvlib.read(antitarget_fname)
-    cnr = commands.do_fix(
-        tgt_raw, anti_raw, ref_probes, diploid_parx_genome=diploid_parx_genome
-    )
-    cns = commands.do_segmentation(
-        cnr, method="cbs", diploid_parx_genome=diploid_parx_genome, threshold=0.001
-    )
-    cll = commands.do_call(cns, diploid_parx_genome=diploid_parx_genome)
-    return (cnr, cns, cll)
-
-
-if __name__ == "__main__":
-    unittest.main(verbosity=2)
+CNVLIB_SUBMODULES = [
+    "access",
+    "antitarget",
+    "autobin",
+    "batch",
+    "bintest",
+    "call",
+    "cluster",
+    "cmdutil",
+    "cnary",
+    "commands",
+    "core",
+    "coverage",
+    "diagram",
+    "export",
+    "fix",
+    "heatmap",
+    "import_rna",
+    "importers",
+    "metrics",
+    "parallel",
+    "params",
+    "plots",
+    "purity",
+    "reference",
+    "reports",
+    "samutil",
+    "scatter",
+    "segfilters",
+    "segmentation",
+    "segmetrics",
+    "smoothing",
+    "vary",
+]
+
+
+@pytest.mark.parametrize("submodule", CNVLIB_SUBMODULES)
+def test_submodule_imports(submodule):
+    importlib.import_module(f"cnvlib.{submodule}")
